LPS-80213 Move Optimize impl to web
