+++ modules/apps/adaptive-media/adaptive-media-web/src/main/java/com/liferay/adaptive/media/web/internal/background/task/OptimizeImagesSingleConfigurationBackgroundTaskExecutor.java
@@ -15,7 +15,7 @@
 package com.liferay.adaptive.media.web.internal.background.task;
 
 import com.liferay.adaptive.media.constants.AMOptimizeImagesBackgroundTaskConstants;
-import com.liferay.adaptive.media.image.optimizer.AMImageOptimizerUtil;
+import com.liferay.adaptive.media.web.internal.optimizer.AMImageOptimizerUtil;
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskExecutor;
 
 /**
